dom/plugins/test/mochitest/test_bug1307694.html

<!DOCTYPE html>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="plugin-utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onLoad="addPerms()">

  <script type="text/javascript">
   SimpleTest.waitForExplicitFinish();
   function addPerms() {
     SpecialPowers.pushPermissions([{type: "plugin:flash", allow: true, context: document}], addPluginElement);
   }

   function addPluginElement() {
     SpecialPowers.DOMWindowUtils.plugins[0].playPlugin();
     var p = document.createElement("embed");
     p.setAttribute("id", "plugin2");
     p.setAttribute("type", "application/x-shockwave-flash-test");
     p.setAttribute("scale", "noscale");
     p.setAttribute("salign", "lt");
     document.body.appendChild(p);
     SimpleTest.executeSoon(function() {
      runTests();
     });
   }

   function runTests() {
     var p = document.getElementById("plugin1");
     ok(p.setColor != undefined, "Static plugin parameter (salign/scale) ordering were correct");
     var p2 = document.getElementById("plugin2");
     ok(p2.setColor != undefined, "Dynamic plugin parameter (salign/scale) ordering were correct");
     SimpleTest.finish();
   }

  </script>
  <p id="display"></p>

  <div id="div1">
    <embed id="plugin1" type="application/x-shockwave-flash-test" width="200" height="200" scale="noscale" salign="lt"></embed>
  </div>
</body>
</html>
